requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
pydantic-settings
alembic
pytest
pytest-asyncio
python-dotenv
httpx
groq
slowapi
